labs/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
PySide6-Essentials==6.6.0
requests==2.31.0
shiboken6==6.6.0
urllib3==2.0.7